move TraceEvent, TraceAction to core
