+++ sk-cli/src/validation/rules/status_field_populated.rs
@@ -1,5 +1,5 @@
+use sk_core::event::TraceEvent;
 use sk_core::prelude::*;
-use sk_store::TraceEvent;
 
 use crate::validation::validator::{
     Diagnostic,
