+++ sk-cli/src/validation/rules/tests/missing_resources_test.rs
@@ -1,6 +1,6 @@
 use assertables::*;
 use serde_json::json;
-use sk_store::TraceEvent;
+use sk_core::event::TraceEvent;
 
 use super::missing_resources::service_account_validator;
 use super::*;
