+++ sk-skel/src/lib.rs
@@ -8,10 +8,8 @@ use std::sync::mpsc;
 
 use pest::Parser;
 use pest_derive::Parser;
-use sk_store::{
-    ExportedTrace,
-    TraceEvent,
-};
+use sk_core::event::TraceEvent;
+use sk_store::ExportedTrace;
 
 use self::ast::{
     Command,
